requirements.txt
#List requirements here

## Data processing
numpy==1.23.5
pandas==1.5.3
opencv-python
PIL


## Modelling
scikit-learn==1.3.1

## visualization
matplotlib
seaborn

# api related things
fastapi         # API framework
uvicorn         # webserver
python-dotenv
requests

# hosting
google-cloud-storage
